create asn file
