+++ core/observables/__init__.py
@@ -2,7 +2,8 @@ from mongoengine import *
 
 from core.observables.tag import ObservableTag, Tag
 from core.observables.observable import Observable
-from core.observables.ip import Ip, AutonomousSystem
+from core.observables.ip import Ip
+from core.observables.asn import AutonomousSystem
 from core.observables.url import Url
 from core.observables.hostname import Hostname
 from core.observables.hash import Hash
