+++ core/observables/ip.py
@@ -52,21 +52,3 @@ class Ip(Observable):
             self.version = 6
 
 
-class AutonomousSystem(Observable):
-
-    """Autonomous System observable"""
-
-    as_num = IntField(verbose_name="Autonomous System number")
-
-    DISPLAY_FIELDS = Observable.DISPLAY_FIELDS + [
-        ("as_num", "Autonomous System number"),
-    ]
-
-    def info(self):
-        info = super(AutonomousSystem, self).info()
-        info["as_num"] = (self.as_num,)
-        return info
-
-    @staticmethod
-    def check_type(txt):
-        return True
